firmware: arm_scmi: add qcom smc xport descriptor
On Qualcomm platforms, we need to modify various parameters:

max_rx_timeout_ms: For PCIe devices, we are performing link-training on
power on request therefore 30ms timeout is not sufficient.
We will refine the value of the timeout after PCI EP is
available to use SCMI solution. For now, we are setting it to 3s.

max_msg_size: It seems that 128 bytes could only hold 12 OPPs on our
platform. Which would mean that agent will need to fetch OPPs in
multiple requests from platform. We want to avoid multiple requests to
save some cycles in boot KPIs hence increasing it to 256 bytes.

max_msg: Keeping the total shmem requirement same, we are reducing it
by half as we doubled the msg_size.

Signed-off-by: Nikunj Kela <nkela@quicinc.com>