spi: geni-qcom: Enable SPI GSI mode for SA8255p Qualcomm platforms
Add support of SPI GSI mode for SA8255p.

In GSI mode, the clock index and divider values must be filled
as part of the TX TRE descriptor. Since we do not have a clock
handler to determine these values, we require a predefined list
of SE DFS frequencies to accurately identify the appropriate index
and divider values.

The list of SE DFS frequencies are 19200000, 32000000, 48000000,
64000000, 96000000, and 100000000

Signed-off-by: Praveen Talari <quic_ptalari@quicinc.com>